debt - move extension debug IPC channel closer to interface
